Dividindo componente Panel em componentes menores
